# backend/main.py
from fastapi import Query
from typing import List
from datetime import datetime
import uvicorn
from backend.database_integration.database import SessionLocal, Base, engine
from backend.database_integration import models as models

# from backend.mistral import extract_event_and_feeling, generate_advice
from sqlalchemy.orm import Session
from fastapi import Depends
from mistralai import Mistral
import os
from fastapi import FastAPI, HTTPException
from dotenv import load_dotenv
from pydantic import BaseModel, Field

load_dotenv()

Base.metadata.create_all(bind=engine)

app = FastAPI(
    title="LifeChat API",
    description="API for logging and analyzing life events and feelings.",
    version="1.0.0",
)


# --- Schemas ---
class Event(BaseModel):
    date: str
    startTime: str
    endTime: str
    description: str
    name: str
    tags: List[str]


class Feeling(BaseModel):
    feelings: List[str]
    score: int = Field(..., ge=1, le=10)
    datetime: str


mistral_client = Mistral(api_key=os.getenv("MISTRAL_API_KEY"))


def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()


@app.get("/")
def read_root():
    return {
        "message": "Welcome to the LifeChat API! Use the endpoints to log events and feelings, and get advice."
    }


@app.post("/addEvent")
def add_event(description: str, db: Session = Depends(get_db)):
    new_event = models.Event(
        startTime=datetime.now(),
        endTime=datetime.now(),
        description=description,
        tags="personal",
    )
    db.add(new_event)
    db.commit()
    db.refresh(new_event)
    return new_event


@app.post("/addFeeling")
def add_feeling(feelings: str, score: int, db: Session = Depends(get_db)):
    new_feeling = models.Feeling(
        feelings=feelings, score=score, datetime=datetime.now()
    )
    db.add(new_feeling)
    db.commit()
    db.refresh(new_feeling)
    return new_feeling


@app.get("/getAllEvents")
def getAllEvents(db: Session = Depends(get_db)):
    events = db.query(models.Event).all()
    return events


@app.get("/getAllFeelings")
def getAllFeelings(db: Session = Depends(get_db)):
    feelings = db.query(models.Feeling).all()
    return feelings


@app.get("/getEvents")
def get_events(
    startTime: datetime = Query(..., description="Start time for the events"),
    endTime: datetime = Query(..., description="End time for the events"),
    db: Session = Depends(get_db),
):
    """
    Get events within a specified date range.
    """
    events = (
        db.query(models.Event)
        .filter(models.Event.startTime >= startTime, models.Event.endTime <= endTime)
        .all()
    )

    if not events:
        raise HTTPException(
            status_code=404, detail="No events found in the specified range."
        )

    return events


@app.get("/getFeelings")
def get_feelings(
    startTime: datetime = Query(..., description="Start time for the feelings"),
    endTime: datetime = Query(..., description="End time for the feelings"),
    db: Session = Depends(get_db),
):
    """
    Get feelings within a specified date range.
    """
    feelings = (
        db.query(models.Feeling)
        .filter(
            models.Feeling.datetime >= startTime, models.Feeling.datetime <= endTime
        )
        .all()
    )

    if not feelings:
        raise HTTPException(
            status_code=404, detail="No feelings found in the specified range."
        )

    return [f.feelings for f in feelings]


@app.get("/getAdvice")
async def get_advice(
    startTime: datetime, endTime: datetime, db: Session = Depends(get_db)
):
    """
    Get advice based on events and feelings in a date range.
    """
    events_db = (
        db.query(models.Event)
        .filter(models.Event.startTime >= startTime, models.Event.endTime <= endTime)
        .all()
    )

    feelings_db = (
        db.query(models.Feeling)
        .filter(
            models.Feeling.datetime >= startTime, models.Feeling.datetime <= endTime
        )
        .all()
    )

    context = f"""
    Events in the period:
    {[event.description for event in events_db]}
    
    Feelings in the period:
    {[feeling.feelings for feeling in feelings_db]}
    """

    chat_response = mistral_client.chat.complete(
        model="mistral-large-latest",
        messages=[
            {
                "role": "system",
                "content": "You are a helpful life coach.Based on the user's events and feelings, provide specific, actionable advice.Keep the advice concise and practical.",
            },
            {"role": "user", "content": context},
        ],
    )

    return chat_response.choices[0].message.content


@app.post("/lifeChat")
async def submit_life_chat(chat: dict, db: Session = Depends(get_db)):
    """
    Process a life chat entry and extract event and feeling information.
    """
    try:
        response = mistral_client.chat.complete(
            model="mistral-large-latest",
            messages=[
                {
                    "role": "system",
                    "content": """Extract event and feeling information from the user's chat.
            Return a JSON with two objects: 'event' and 'feeling'.
            Event should include date, startTime, endTime, description, and tags.
            Feeling should include feelings (list of emotions), score (1-10), and datetime.""",
                },
                {"role": "user", "content": chat["chat"]},
            ],
        )

        extracted_data = eval(response.choices[0].message.content)

        event = Event(**extracted_data["event"])
        feeling = Feeling(**extracted_data["feeling"])

        db.add(event)
        db.add(feeling)
        db.commit()
        db.refresh(event)

        return {"event": event, "feeling": feeling}

    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))


if __name__ == "__main__":
    uvicorn.run("main:app", host="127.0.0.1", port=8000, reload=True)